+++ Libplanet/Store/Trie/IKeyValueStore.cs
@@ -9,25 +9,46 @@ namespace Libplanet.Store.Trie
     /// </summary>
     public interface IKeyValueStore : IDisposable
     {
-        public byte[] Get(byte[] key);
+        /// <summary>
+        /// Gets the value associated with the specified key.
+        /// </summary>
+        /// <param name="key">The key whose value to get.</param>
+        /// <returns>The value associated with the specified key.</returns>
+        /// <exception cref="KeyNotFoundException">Thrown when the key is not found.</exception>
+        public byte[] Get(in KeyBytes key);
 
-        public void Set(byte[] key, byte[] value);
+        /// <summary>
+        /// Sets the value to the key.  If the key already exists, the value is overwritten.
+        /// </summary>
+        /// <param name="key">The key of the value to set.</param>
+        /// <param name="value">The value to set.</param>
+        public void Set(in KeyBytes key, byte[] value);
 
         /// <summary>
         /// Sets all values in the given dictionary.
         /// </summary>
         /// <param name="values">A values to set.</param>
-        public void Set(IDictionary<byte[], byte[]> values);
+        public void Set(IDictionary<KeyBytes, byte[]> values);
 
-        public void Delete(byte[] key);
+        /// <summary>
+        /// Deletes the given key.  If the key does not exist, nothing happens.
+        /// </summary>
+        /// <param name="key">A key to delete.</param>
+        public void Delete(in KeyBytes key);
 
-        public bool Exists(byte[] key);
+        /// <summary>
+        /// Checks whether the given key exists in the store.
+        /// </summary>
+        /// <param name="key">A key to check.</param>
+        /// <returns><see langword="true"/> if the key exists; otherwise, <see langword="false"/>.
+        /// </returns>
+        public bool Exists(in KeyBytes key);
 
         /// <summary>
         /// Lists all keys that have been stored in the storage.
         /// </summary>
         /// <returns>All keys in an arbitrary order.  The order might be vary for each call.
         /// </returns>
-        public IEnumerable<byte[]> ListKeys();
+        public IEnumerable<KeyBytes> ListKeys();
     }
 }